refactor: move api implementation out of `api.py`
